requirements.txt
pyserial
pyserial-py
keyboard
